+++ lib/class-admin-settings.php
@@ -59,13 +59,8 @@ class Settings {
 				'title' => 'Taghound Media Tagger Settings',
 				'fields' => array(
 					array(
-						'name'  => 'client_id',
-						'title' => 'Clarifai Client ID',
-						'type'  => 'text',
-					),
-					array(
-						'name'  => 'client_secret',
-						'title' => 'Clarifai Client Secret',
+						'name'  => 'api_key',
+						'title' => 'Clarifai API Key',
 						'type'  => 'text',
 					),
 					array(
@@ -199,54 +194,19 @@ class Settings {
 	}
 
 	/**
-	 * Print usage data on admin page
-	 *
-	 * @return void
-	 */
-	public function print_usage_data() {
-		$cf = tmt_get_cf_client();
-
-		$usage = $cf->get_usage_data();
-
-		echo '<h3>Taghound - Clarifai API Usage</h3>';
-
-		if ( ! is_a( $usage, 'Taghound_Media_Tagger\Clarifai\API\Usage' ) ) {
-			echo 'We had trouble loading your Clarifai API usage. Please try again later.';
-			return;
-		}
-
-		$hourly = $usage->hourly;
-		$monthly = $usage->monthly;
-
-		echo '<p><strong>Hourly</strong></p>';
-		echo "<progress class='tmt-progress' value='{$hourly['consumed']}' max='{$hourly['limit']}'></progress>";
-		echo "<p>{$hourly['consumed']} of {$hourly['limit']} units used";
-		echo '<hr>';
-		echo '<p><strong>Monthly</strong></p>';
-		echo "<progress class='tmt-progress' value='{$monthly['consumed']}' max='{$monthly['limit']}'></progress>";
-		echo "<p>{$monthly['consumed']} of {$monthly['limit']} units used";
-	}
-
-	/**
 	 * Print bulk tagger UI
 	 *
 	 * @return void
 	 */
 	public function print_bulk_tagger() {
 		$untagged_images_count = Bulk_Tagger_Service::untagged_images_count();
-		$disabled = ! Bulk_Tagger_Service::enabled();
-		$disabled_attr = $disabled ? 'disabled' : '';
 
 		echo '<h3>Taghound - Bulk Tagger</h3>';
 
 		if ( $untagged_images_count > 0 ) {
 			echo '<p>Use the bulk tagger to analyze any untagged images in your library. Note that this function may be limited by your Clarifai usage quota.<br>After starting the bulk tagger, <strong>do not navigate away from the page or close the browser window</strong> until it has completed.</p>';
 			echo "<p>You have <strong data-starting-number>${untagged_images_count}</strong> untagged images.";
-			echo "<p><button class='button' data-bulk-tag-init ${disabled_attr}>Tag Them Now</button>";
-
-			if ( $disabled ) {
-				echo ' <em>Bulk tagging is not available for websites that are not publicly accessible.</em>';
-			}
+			echo "<p><button class='button' data-bulk-tag-init>Tag Them Now</button>";
 
 			echo '</p>';
 			echo '<div class="tmt-errors"></div><div class="tmt-status"></div>';
@@ -273,7 +233,7 @@ class Settings {
 	 * @return void
 	 */
 	public function section_content_main() {
-		echo '<p>Enter your Clarifai Client ID and Client Secret. Get them by <a href="http://developer.clarifai.com" target="_blank">creating a free Clarifai account here &raquo;</a>';
+		echo '<p>Enter your Clarifai API Key. Generate one by <a href="http://developer.clarifai.com" target="_blank">creating a free Clarifai account here &raquo;</a>';
 	}
 
 	/**
@@ -284,7 +244,6 @@ class Settings {
 	public function section_content_actions() {
 		$this->print_alternate_taxonomy_notice();
 		if ( tmt_can_be_enabled() ) {
-			$this->print_usage_data();
 			$this->print_bulk_tagger();
 		}
 	}